Moved Config.java to com.teragrep.cfe_39.configuration package for further refactoring.
